Insert geocoding task into gtr pipeline.
